fix(canon): preserve undefined withDefaults values
